+++ src/trx/game/objects/types.h
@@ -47,7 +47,6 @@ typedef struct {
     bool ignore_rot;
 } OBJECT_BOUNDS;
 
-typedef struct SAVEGAME_LEGACY_IO SAVEGAME_LEGACY_IO;
 typedef struct SG_READ_IO SG_READ_IO;
 typedef struct SG_WRITE_IO SG_WRITE_IO;
 
@@ -73,7 +72,6 @@ typedef struct OBJECT {
     void (*handle_save_func)(ITEM *item, SAVEGAME_STAGE stage);
     void (*priv_load_func)(ITEM *item, SG_READ_IO *io);
     void (*priv_save_func)(const ITEM *item, SG_WRITE_IO *io);
-    void (*priv_legacy_load_func)(ITEM *item, const SAVEGAME_LEGACY_IO *io);
     const OBJECT_BOUNDS *(*bounds_func)(void);
     bool (*is_usable_func)(int16_t item_num);
     void (*add_walkable_func)(int16_t item_num);