+++ docs/CHANGELOG.md
@@ -51,7 +51,7 @@
     - `on_control_post` → `after_control`
 - changed turbo cheat to auto‑reset to normal speed if pushed past limit, making it easier for new players to recover from accidental changes
 - changed Blades to support being reset
-- removed support for < TRX 1.0 savegames
+- removed support for legacy (TombATI / TR2 GOG/Steam) and pre-1.0 (TR1X/TR2X) savegame files
 - fixed drawing debug triggers using random tint near water sources
 - fixed drawing debug triggers glitching through triangular portals
 - fixed drawing shadows twice when item intersects a portal (#4640, regression from 1.0)